requirements.txt
tkinter
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

